perf(loader): run reintern_directives on fresh-parse path + drop dead SyncStringInterner (closes #1071)
The cache-hit path already ran reintern_directives after deserializing
to deduplicate InternedStr across files; the fresh-parse path did
not. So multi-file ledgers parsed from source had identical account /
currency / tag strings landing in distinct Arc<str> allocations
(per-file StringInterner), defeating the Arc::ptr_eq fast path in
InternedStr::PartialEq and forcing every cross-file equality through
byte comparison.

Changes:

(1) Extract reintern_directives and reintern_plain_directives from
the feature-gated cache module into a new always-on dedup module.
The dedup walk has no rkyv/cache dependency; gating it behind the
cache feature was incidental. lib.rs now re-exports both functions
unconditionally.

(2) Call dedup::reintern_directives at the end of Loader::load, so
every consumer of LoadResult sees a deduplicated directive list
regardless of whether they came in via cache or fresh parse.

(3) Drop SyncStringInterner from rustledger-core::intern. The type
had zero call sites in the workspace; an audit for #1071 confirmed
it was dead code. Removing it also lets us drop the parking_lot
dependency from rustledger-core (added in #1076 specifically for
this type).

Tests:
* New test_fresh_parse_deduplicates_internedstr_across_files locks
  in the cross-file dedup invariant via VirtualFileSystem (two files
  referencing Assets:Bank; assert the resulting Posting and Open
  share an Arc via InternedStr::ptr_eq).
* Existing reintern unit tests in cache.rs continue to pass against
  the moved implementation.

Closes #1071
