+++ crates/rustledger-core/Cargo.toml
@@ -29,7 +29,6 @@ serde.workspace = true
 serde_json.workspace = true
 rkyv = { workspace = true, optional = true }
 rustc-hash.workspace = true
-parking_lot.workspace = true
 smallvec.workspace = true
 
 [dev-dependencies]
